fixup! MAILBOX-314 add a Integration test for search with ES without TIKA
