+++ mailbox/elasticsearch/src/main/java/org/apache/james/mailbox/elasticsearch/events/ElasticSearchListeningMessageSearchIndex.java
@@ -79,12 +79,17 @@ public class ElasticSearchListeningMessageSearchIndex extends ListeningMessageSe
 
     @Override
     public EnumSet<SearchCapabilities> getSupportedCapabilities(EnumSet<MessageCapabilities> messageCapabilities) {
-        return EnumSet.of(
+        EnumSet<SearchCapabilities> supportedCapabilites = EnumSet.of(
             SearchCapabilities.MultimailboxSearch,
             SearchCapabilities.Text,
             SearchCapabilities.FullText,
-            SearchCapabilities.Attachment,
             SearchCapabilities.PartialEmailMatch);
+
+        if (messageToElasticSearchJson.handleIndexAttachment()) {
+            supportedCapabilites.add(SearchCapabilities.Attachment);
+        }
+
+        return supportedCapabilites;
     }
     
     @Override
